Unify custom attribute values check
